requirements.txt
fastapi==0.86.0
keybert==0.6.0
multi_rake==0.0.2
numpy==1.21.6
pandas==1.4.2
plotly==5.8.0
pydantic==1.9.1
pytextrank==3.2.4
python-dotenv==0.21.0
requests==2.28.1
spacy==3.4.2
uvicorn==0.17.6
